VideoFramesSceneExtractor-lambda/requirements.txt
boto3>=1.34.0
opencv-python-headless>=4.8.0
scenedetect>=0.6.2
supabase>=2.0.0
python-dotenv>=1.0.1
